move iterator utils to common directory
